+++ lib/src/experimental/pimlico/pimlico.dart
@@ -1,6 +1,0 @@
-/// Experimental Pimlico utilities.
-///
-/// **Warning:** These are experimental and the API may change.
-library;
-
-export 'erc20_paymaster.dart';