+++ packages/permissionless_passkeys/example/lib/screens/account_screen.dart
@@ -127,10 +127,9 @@ class _AccountScreenState extends State<AccountScreen> {
                   children: [
                     Text(
                       'Send Transaction',
-                      style:
-                          Theme.of(context).textTheme.titleMedium?.copyWith(
-                                fontWeight: FontWeight.bold,
-                              ),
+                      style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                            fontWeight: FontWeight.bold,
+                          ),
                     ),
                     const SizedBox(height: 4),
                     Text(